requirements.txt
Flask~=2.2.2
flask_session~=0.4.0
psycopg2-binary~=2.9.3
flask_bcrypt~=1.0.1